FastApi/requirements.txt
docker
markdown2
openpyxl
python-dotenv
kaleido==0.2.1
plotly==5.3.1
langchain-openai==0.1.24
langchain-community==0.2.17
langchain-cohere==0.1.9
langchain==0.2.16
xlrd
traceback2
pandas
uvicorn
numpy
aiofiles
crewai
crewai_tools
python-multipart
reportlab==4.2.5